test: Sync handshake register handling with upstream
Access to the sensor registers is affected during a secure update. The
BMC firmware becomes unavailable to respond to register requests. During
testing in preparation for upstreaming, it was discovered that previous
implementation was not correct. Access to the sensor register during
the write phase of an update _are_ affected, but access during the
authentication/programming phase is not. These changes align the backport
driver with changes that will be submitted upstream.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>